Rename files properly
