+++ LAB1_atlast/model.py
